% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PLSDA}
\alias{summary.PLSDA}
\title{summary.PLSDA,
This function prints the standardized projection coefficients of matrix X}
\usage{
\method{summary}{PLSDA}(object, ...)
}
\arguments{
\item{a}{plsda object}
}
\value{
prints the standardized (centered and reduce) projection coefficients  of matrix X

prints the restitution quality of the pls regression
}
\description{
summary.PLSDA,
This function prints the standardized projection coefficients of matrix X
}
\examples{
data(iris)
formule = Species~.

PLSDA_object = fit(formula=formule, data=iris)
summary(PLSDA_object)

}
